+++ index.js
@@ -7,35 +7,66 @@
 const Bone = require('./lib/bone');
 const Collection = require('./lib/collection');
 const connect = require('./lib/connect');
-const DataTypes = require('./lib/data_types');
+const DataTypes =require('./lib/data_types');
 const migrations = require('./lib/migrations');
+const sequelize = require('./lib/adapters/sequelize');
 
 class Realm {
-  constructor(opts) {
-    // setup the default model to cope with legacy api
-    this.Model = Bone;
-    this.options = {
-      client: 'mysql',
-      database: opts.db,
-      ...opts,
-      define: { underscored: false, ...opts.define },
+  constructor(opts = {}) {
+    const { client, dialect, database, Bone: Osteon, ...restOpts } = {
+      client: opts.dialect || 'mysql',
+      database: opts.db || opts.storage,
+      Bone: class Spine extends Bone {},
+      ...opts
     };
+    const Spine = dialect ? sequelize(Osteon) : Osteon;
+
+    this.Bone = Spine;
+    this.models = Spine.models = {};
+    this.options = Spine.options = {
+      client,
+      dialect,
+      database,
+      ...restOpts,
+      define: { underscored: true, ...opts.define },
+    };
+
+    const { define } = this.options;
+    for (const prop of [ 'createdAt', 'updatedAt', 'deletedAt' ]) {
+      if (!define.hasOwnProperty(prop)) define[prop] = prop;
+    }
   }
 
   get driver() {
-    const { driver } = this.Model;
+    const { driver } = this.Bone;
     if (!driver) throw new Error('database not connected yet');
     return driver;
   }
 
-  async connect() {
-    const Model = await connect(this.options);
-    this.Model = Model;
+  define(name, attributes, opts = {}) {
+    const { Bone } = this;
+    const Model = class extends Bone {};
+    Object.defineProperty(Model, 'name', { value: name });
+    Model.init(attributes, opts);
+    Bone.models[name] = Model;
     return Model;
+  }
+
+  async connect() {
+    await connect({ ...this.options, Bone: this.Bone });
+  }
+
+  async sync() {
+    const { Bone } = this;
+    if (!Bone.driver) await this.connect();
+
+    for (const model of Object.values(Bone.models)) {
+      await model.sync();
+    }
   }
 }
 
-Object.assign(Realm.prototype, { ...migrations, DataTypes });
-Object.assign(Realm, { connect, Bone, Collection, DataTypes });
+Object.assign(Realm.prototype, migrations, { DataTypes });
+Object.assign(Realm, { connect, Bone, Collection, DataTypes, sequelize });
 
 module.exports = Realm;