+++ test/integration/suite/data_types.js
@@ -1,10 +1,11 @@
 'use strict';
 
 const assert = require('assert').strict;
-const { Bone } = require('../../..');
-const { INTEGER, STRING, DATE, TEXT } = Bone;
+const { Bone, DataTypes } = require('../../..');
+const { INTEGER, STRING, DATE, TEXT } = DataTypes;
 
-const Note = Bone.define('Note', {
+class Note extends Bone {};
+Note.init({
   id: INTEGER,
   title: STRING,
   body: TEXT,
@@ -18,6 +19,7 @@ describe('=> Data Types', () => {
       columnName: 'title',
       dataType: STRING,
       defaultValue: null,
+      type: String,
     });
   });