+++ test/unit/drivers/test.sqlite.js
@@ -1,7 +1,10 @@
 'use strict';
 
 const assert = require('assert').strict;
+const { DataTypes } = require('../../..');
 const SqliteDriver = require('../../../lib/drivers/sqlite');
+
+const { INTEGER } = DataTypes;
 
 const driver = new SqliteDriver('sqlite', {
   database: '/tmp/leoric.sqlite3',
@@ -18,5 +21,13 @@ describe('=> SQLite driver', () => {
     for (const column of columns) {
       for (const prop of props) assert.ok(column.hasOwnProperty(prop));
     }
+  });
+
+  it('driver.createTable(table, definitions)', async () => {
+    await driver.dropTable('notes');
+    await driver.createTable('notes', {
+      id: { dataType: INTEGER, primaryKey: true, autoIncrement: true },
+      public: { dataType: INTEGER },
+    });
   });
 });